- moved profile stock generation into load_profile - replaced some np.divide by / - fixed cli and run.py
